Changes 20-05

+++ src/modelo/negocio/DadosTecMensais.java
@@ -20,7 +20,7 @@ public class DadosTecMensais extends DatabaseObject implements Serializable {
 
     private int mes;
     private int ano;
-    private int dado;
+    private double dado;
     private Indicador indicador; // indicador associado a cada dado tecnico mensal
     private Perfil perfil;
     
@@ -29,7 +29,7 @@ public class DadosTecMensais extends DatabaseObject implements Serializable {
     }
 
 
-    public DadosTecMensais(int mes, int ano, int dado, Indicador indicador, Perfil perfil) {
+    public DadosTecMensais(int mes, int ano, double dado, Indicador indicador, Perfil perfil) {
         super("dados_tecnicos_mensais", "idDTM");
         
         this.mes = mes;
@@ -55,11 +55,11 @@ public class DadosTecMensais extends DatabaseObject implements Serializable {
         this.ano = ano;
     }
 
-    public int getDado() {
+    public double getDado() {
         return dado;
     }
 
-    public void setDado(int dado) {
+    public void setDado(double dado) {
         this.dado = dado;
     }
 
@@ -96,7 +96,7 @@ public class DadosTecMensais extends DatabaseObject implements Serializable {
     public void setObjectData(Map<String, Object> data) {
         mes = Cast.toInt(data.get("mes"));
         ano = Cast.toInt(data.get("ano"));
-        dado = Cast.toInt(data.get("dado"));
+        dado = Cast.toDouble(data.get("dado"));
         indicador = new GenericDAO<>(Indicador.class).retrieve(Cast.toInt(data.get("idDTM_indicadorFK")));
         perfil =  new GenericDAO<>(Perfil.class).retrieve( Cast.toInt(data.get("idPerfilFK")));
     }
